Create _exec.model module
